[Update/Fix]params emptyのエラーを解消

+++ app/controllers/public/orders_controller.rb
@@ -5,7 +5,12 @@ class Public::OrdersController < ApplicationController
 
   def confirm #要確認
     @order = Order.new(order_params)
-    @new_order = Order.new
+    @order.customer_id = current_customer.id
+    @order.payment_type = params[:order][:payment_type]
+    @order.postage = 800
+    @cart_items = current_customer.cart_items.all
+    @total_amount = @cart_items.sum(&:subtotal)
+    @order.total_payment = @total_amount
     if params[:order][:select_address] == "1"
       @order.postal_code = current_customer.postal_code
       @order.address = current_customer.address
@@ -22,10 +27,6 @@ class Public::OrdersController < ApplicationController
     # else
     #   render 'new'
     end
-    @order.postage = 800
-    @cart_items = current_customer.cart_items.all
-    @total_amount = @cart_items.sum(&:subtotal)
-    @order.total_payment = @total_amount
   end
 
   def thanks
@@ -33,20 +34,21 @@ class Public::OrdersController < ApplicationController
 
   def create
     @order = Order.new(order_params)
-    #@order.customer_id = current_customer.id
-    @order.save
-
-    current_customer.cart_items.each do |cart_item|
-      @order_details = OrderDetails.new
-      @order_details.item_id = cart_item.item_id
-      @order_details.amount = cart_item.amount
-      @order_details.price_tax_in = (cart_item.item.price*1.10).floor
-      @order_details.order_id = @order.id
-      @order_details.save
+    @order.customer_id = current_customer.id
+    if @order.save
+      current_customer.cart_items.each do |cart_item|
+        @order_details = OrderDetail.new
+        @order_details.item_id = cart_item.item_id
+        @order_details.amount = cart_item.amount
+        @order_details.price_tax_in = cart_item.item.price*1.10
+        @order_details.order_id = @order.id
+        @order_details.save
+      end
+      current_customer.cart_items.destroy_all
+      redirect_to thanks_path
+    else
+      render "new"
     end
-
-    current_customer.cart_items.destroy_all
-    redirect_to thanks_path
   end
 
   def index
@@ -60,6 +62,6 @@ class Public::OrdersController < ApplicationController
   private
 
   def order_params
-    params.permit(:payment_type, :address, :postal_code, :receiver_name, :postage, :total_payment, :status, :customer_id)
+    params.require(:order).permit(:payment_type, :address, :postal_code, :receiver_name, :postage, :total_payment, :status, :customer_id)
   end
 end
